refactor: step 생성 로직 분리
