+++ src/main/java/model/Layer.java
@@ -1,20 +1,11 @@
 package model;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class Layer {
     private final List<Step> steps;
 
-    public Layer(int numberOfParticipants) {
-        List<Step> steps = new ArrayList<>();
-        boolean priorExistence = false;
-
-        for (int i = 1; i < numberOfParticipants; i++) {
-            boolean doesExist = StepExistenceGenerator.generate(priorExistence);
-            priorExistence = doesExist;
-            steps.add(Step.findByExistence(doesExist));
-        }
+    public Layer(List<Step> steps) {
         this.steps = steps;
     }
